refactor: split test_cli_workspace_commands.py into focused modules
Split the 3118-line test_cli_workspace_commands.py into 5 smaller files:
- test_cli_workspace_merge.py (768 lines) - merge/review/discard/preview commands
- test_cli_workspace_pr.py (417 lines) - PR creation commands
- test_cli_workspace_conflict.py (740 lines) - conflict detection functions
- test_cli_workspace_worktree.py (516 lines) - worktree management commands
- test_cli_workspace_utils.py (1449 lines) - utilities and edge cases

Also:
- Created test_utils.py with shared configure_build_mocks helper
- Updated 7 tests in test_cli_build_commands.py to use configure_build_mocks
- Removed permanently-skipped test

This improves test discoverability, reduces file sizes, and makes the test
suite more maintainable while preserving all test coverage.
